+++ src/platformsupport/fontdatabases/mac/qfontengine_coretext.mm
@@ -39,9 +39,11 @@
 
 #include <private/qimage_p.h>
 
+#include <cmath>
+
 QT_BEGIN_NAMESPACE
 
-static float SYNTHETIC_ITALIC_SKEW = tanf(14 * acosf(0) / 90);
+static float SYNTHETIC_ITALIC_SKEW = std::tan(14.f * std::acos(0.f) / 90.f);
 
 bool QCoreTextFontEngine::ct_getSfntTable(void *user_data, uint tag, uchar *buffer, uint *length)
 {
